requirements.txt
Flask>=2.3.0
Flask-SQLAlchemy>=3.0.0
Flask-Login>=0.6.0
Flask-WTF>=1.1.0
WTForms>=3.0.0
numpy>=1.26.0
pandas>=2.1.0
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.17.0
gunicorn>=21.2.0
python-dotenv>=1.0.0
Werkzeug>=2.3.0
Jinja2>=3.1.0
MarkupSafe>=2.1.0
itsdangerous>=2.1.0
click>=8.1.0
blinker>=1.6.0
Flask-OAuthlib>=0.9.6
requests-oauthlib>=1.3.1
PyJWT>=2.8.0
cryptography>=41.0.0
